Add files via upload
Python code to obtain metrics like receiver operating characteristics (ROC) curve and area under the curve (AUC) from scratch without using in-built functions.
Algorithm for calculating the roc curve and calculating the threshold for a given probability of correct recognition. The algorithm is optimized to reduce the consumption of operative memory.